examples/mode/token_creation/requirements.txt
goat-sdk>=0.1.0
python-dotenv>=1.0.0
web3>=6.11.3
